balance changes

+++ Assets/Scripts/Player/Sword.cs
@@ -20,7 +20,7 @@ public class Sword : MonoBehaviour
     {
         anim = FindObjectOfType<Animator>();
         hitReg = transform.parent;
-        damage = (int)((float)damage * Buffs.SwordDamage);
+        damage = (int)(damage * Buffs.SwordDamage);
     }
 
     private void Update()
